requirements.txt
# Core data processing libraries
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning libraries (essential for data quality analysis)
scikit-learn>=1.3.0
scipy>=1.10.0

# Data Quality Analysis Tools
ydata-profiling>=4.5.0

# Statistical analysis
statsmodels>=0.14.0

# Data validation and cleaning
cleanlab>=2.4.0
pydantic>=2.0.0

# Outlier detection (used by your script)
pyod>=1.1.0

# Additional ML libraries for advanced analysis
xgboost>=1.7.0
lightgbm>=4.0.0

# Data drift detection
evidently>=0.4.0

# Great Expectations for data validation
great-expectations>=0.18.0

# Additional utilities
requests>=2.31.0
python-dateutil>=2.8.0
jsonschema>=4.17.0

# Plotting and visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress bars
tqdm>=4.64.0

# Configuration management
python-dotenv>=1.0.0

# File handling
openpyxl>=3.1.0

# JSON handling
orjson>=3.8.0

# Performance optimization
numba>=0.57.0

# Memory profiling
memory-profiler>=0.60.0

# Model interpretability
shap>=0.42.0

# Feature engineering
feature-engine>=1.6.0

# Advanced statistics
pingouin>=0.5.0

# Time series analysis
prophet>=1.1.0

# Automated ML
auto-sklearn>=0.15.0

# Deep learning for anomaly detection
tensorflow>=2.13.0

# Advanced clustering
hdbscan>=0.8.0

# Imbalanced learning
imbalanced-learn>=0.11.0

# Hyperparameter optimization
optuna>=3.0.0

# Model validation
yellowbrick>=1.5.0